cryptohome: Add missing lock when accessing mounts
The cryptohome mounts can be accessed concurrently by multiple threads
and have a lock (mounts_lock_) that must be held before they can be
accessed.  Unfortunately, when iterating over the mounts after
successfully initializing the TPM, the Service::InitializeTpmComplete
function did not acquire the lock.

Since the time this bug was introduced (commit b009d07 in March, 2013)
the tpm initialization flow has apparently been fast enough that it
never overlapped with the session teardown phase of the login_Cryptohome
test.  This changed recently and the race condition is frequently
triggered.

Add the missing lock acquisition around the mount iteration code in
Service::InitializeTpmComplete.

BUG=chromium:616854
TEST=login_Cryptohome on peach_pit

Change-Id: I8ab4e4a6e5b17beb341cb86b13dc0dc258e04f19
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/362609
Reviewed-by: Andrey Pronin <apronin@chromium.org>