power: multiply power table value by 8 for Intel driver
Intel driver expects the power table value coming from user to be 8
times of the value measured by the instrument. To avoid confusion and
align with the Intel's iw fork, we expect the value coming from user to
be direct result of the measurement.

BUG=chrome-os-partner:51246
TEST=`emerge-cave power_manager`

Change-Id: I6f369535d934d60777b1aeddb406d3f589827878
Reviewed-on: https://chromium-review.googlesource.com/373318
Commit-Ready: Wei-Ning Huang <wnhuang@chromium.org>
Tested-by: Wei-Ning Huang <wnhuang@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>