shill: Add reorder-services.sh script
This script tells shill to prefer wifi over ethernet, and it creates
a route back to the host PC so that ssh and adb continue to work.  It
is primarily intended for use with CTS, because some CTS tests are most
useful when the default network interface is wifi rather than ethernet
(but an ethernet control path is always needed to invoke CTS).

BUG=b:30471867
TEST=run reorder-services.sh by hand

Change-Id: Icd5b8477a9bd921a2c38585d2a1619ccfbd83c05
Reviewed-on: https://chromium-review.googlesource.com/373072
Commit-Ready: Kevin Cernekee <cernekee@chromium.org>
Tested-by: Kevin Cernekee <cernekee@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@google.com>
Reviewed-by: Kirtika Ruchandani <kirtika@google.com>