crosh: add support for module dirs
We want to let boards add their own sets of functions to crosh w/out
having to modify crosh itself.  Add a set of directories that we will
load blindly.

We don't want this to become a generic mechanism for CrOS for adding
code to crosh due to security/review concerns.  We'll enforce that
via a new autotest.

BUG=chromium:640401
TEST=running crosh locally loads local fragments dynamically

Change-Id: I77393f2ff5cbd32cb746a0ab8c302d85d01f202a
Reviewed-on: https://chromium-review.googlesource.com/378776
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>