attestation: fix waiting until idle for testing
AttestationService::WaitUntilIdleForTesting returned too early when
there were no tasks in the incoming queue, but the current task
was still running. That's caused by the fact that MessageLoop::
IsIdleForTesting checks on the incoming queue. Fixed by posting a
task to the incoming queue that checks and signals if the message
loop is idle when it runs, i.e. the incoming queue is empty + no
tasks run since the previous invocation of that task.

BUG=chrome-os-partner:56862
TEST=add 1 sec delay to PrepareForEnrollment, check that unit test
     AttestationServiceTest.PrepareForEnrollment still passes.

Change-Id: I14eb453b6519dcbb228fdf5d7b9ec2548d9219fa
Reviewed-on: https://chromium-review.googlesource.com/380379
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>