power: memory_suspend_test: Use 2.7 GB as fallback, not 2.9 GB
It appears that 2.9 GB isn't quite guaranteed to pass in 32-bit
userspace on all systems.  Let's be a little more conservative and go
down to 2.7 GB.  If there are really memory errors this will be almost
as likely to catch them anyway and then we don't have to try to fiddle
more.

Note: from testing in python, the max I could allocate on rk3399-kevin
was about 0xab175000 bytes.  Specifically:

  >>> import ctypes
  >>> libc = ctypes.CDLL("libc.so.6")
  >>> p = libc.malloc(0xab175000); p; libc.free(p)
  32776
  0
  >>> p = libc.malloc(0xab176000); p; libc.free(p)
  0
  0

From above, max:    2,870,431,744 bytes.
2800 * 1024 * 1024: 2,936,012,800 bytes.
2700 * 1024 * 1024: 2,831,155,200 bytes.

NOTE: it's possible (probable) that we can actually go higher than
0xab175000 bytes in C code (because of the missing python overhead), but
let's just be safe and stick with something that should definitely work.

BUG=chromium:638755
TEST=suspend_stress_test --memory_check no longer crashes.

Change-Id: I600eceec1d5be203ef36e288f8a80c5a838ee49a
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/382151
Reviewed-by: Dan Erat <derat@chromium.org>