cryptohome: fix race for InitializeTpmComplete and mount_thread
In Service::Initialize(), tpm initialization that runs in a parallel
thread is started before a call to mount_thread_.Start(). When the tpm
initialization completes, it posts a task to a MessageLoop of mount_thread_.
If mount_thread_ is not started by then, we get a null pointer dereference.

BUG=b:31371846
TEST=unittests (new added: ServiceTestNotInitialized.CheckTpmInitRace)

Change-Id: I13765dc8588661a949ed214e0d7c5c9b5a55aa4a
Reviewed-on: https://chromium-review.googlesource.com/383233
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>