login: Add EmitArcBooted D-Bus method
to allow Chrome to send arc-booted Upstart signal.

Currently, the signal is sent by /etc/init/arc-boot-watcher.conf
like this:

 script
   [...]
   while ! /usr/sbin/android-sh -c 'getprop sys.boot_completed' | grep 1 ;
   do
     sleep 1
   done
   initctl emit arc-booted
 end script

but as you can see, it's based on a polling and is not very clean.
With this CL, we can remove the Upstart hack.

BUG=b:31254747
TEST=cros_run_unit_tests --board=cyan --packages chromeos-login
TEST=cros_run_unit_tests --board=rambi --packages chromeos-login
TEST='grep arc-booted /var/log/messages' on a device
CQ-DEPEND=CL:384616

Change-Id: I655acbf8783f16d768cc24e36a7a9e68f1b69b41
Reviewed-on: https://chromium-review.googlesource.com/384594
Commit-Ready: Yusuke Sato <yusukes@chromium.org>
Tested-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>