login: ui.conf - Place ui processes in a cpu cgroup
This moves all of the Chrome processes to the /ui cpu cgroup.  Moving
these processes to a cgroup at a parallel level to Android allows for
setting relative priority between Chrome and Android by setting the
relative cpu.shares of each cgroup.

Chrome renderers will still be put in the top level and will be
unaffected by this change, however a Chrome change will move them down
to be children of Chrome so they can be boosted/cut as well.

BUG=chrome-os-partner:56243
TEST=check system responsiveness after login

Signed-off-by: Dylan Reid <dgreid@chromium.org>
Change-Id: I2c391abc665e4f8d98c17a06701d80b7b54e4d49
Reviewed-on: https://chromium-review.googlesource.com/380327
Reviewed-by: David Reveman <reveman@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>