biod: fix closed fake biometric pipe from live locking the message loop
The named pipe used by FakeBiometric for input is added to the message
loop in order for it to get signaled when there is input. Due to the way
pipes work, a poll on their FD will always trigger a POLLHUP after the pipe is
opened writable and closed, which fake_biometric_tool does. Because this
constantly happens, no other events could make it through the message
loop. The workaround is to open the pipe read/write in FakeBiometric to
ensure a writable end is always open, which prevents POLLHUP.

BUG=None
TEST=send data using fake_biometric_tool and then send something via
     dbus-send and ensure a response comes without timing out

Change-Id: I1552f3c93fc845c76d7e859baad1ec95cf0a36c9
Reviewed-on: https://chromium-review.googlesource.com/384638
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>