Add command line mount option for system boot up.
Flash player will be mounted at boot time in a fixed location, so this
CL adds the command line options to allow imageloader to be executed in
this manner. It also deletes the LoadComponent dbus interface, which
will be re-added when other clients are ready to use it.

BUG=chromium:630421
TEST=emerge-${BOARD} imageloader and run on device

Change-Id: Ie562d9aecb1031f45bce2f0b64ec4e6dd8b3181a
Reviewed-on: https://chromium-review.googlesource.com/385087
Commit-Ready: Greg Kerr <kerrnel@chromium.org>
Tested-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Ricky Zhou <rickyz@chromium.org>