libcontainer: add cpuset cgroup to the container
Add cpuset to the list of cgroups that are available to the container.

Note that cpuset is a bit special: we need to copy the "cpus" and "mems"
from parent cpuset (if it exists), otherwise we'll start with an
unusable (empty) cpuset, since nothing can be moved or started in it.
Also, because available CPUs and memory nodes are system-specific, we
can't do what we do for scheduling groups and rely on config files to
set them up, as configs are usually shared between systems.

BUG=chromium:647994
TEST=start android container, verify cpuset cgroup is present there.

Change-Id: Ib52ca2db8c33afd4290e3e2be3aa168bf62492f0
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/386792
Reviewed-by: Dylan Reid <dgreid@chromium.org>