libchromeos-ui: chromium_command_builder: add flag for passive event listeners
This CL adds a bit of logic to chromium command builder to add the
"--passive-listeners-default=true" flag if the correct USE flag was set
for this board.  The USE flag is "passive_event_listeners".  That that
USE flag is found in /etc/ui_use_flags.txt (set based on the USE flags
for that board during build time) when starting Chrome, then
--passive-listeners-default=true will be added to the Chrome command
line.

CQ-DEPEND=CL:387745,CL:387726
BUG=chrome-os-partner:57500
TEST=I emerged and deployed libchromeos-ui onto my Kevin.  By looking at
chrome://version I could see the command line and check that the right
flag had been added.  When I retried after removing the USE flag, the
command line when back to normal.  I followed up using
rbyers.net/eventTest.html and "inspect"ed the main area's touchstart
event listener.  This correctly identified the listeners as passive or
not passive as the case may be for my tests.

Change-Id: Icbf989d2e856935dc536ebf4e29fdd00d87b2578
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/387747
Reviewed-by: Dan Erat <derat@chromium.org>