cros-disks: improve handling of unmounting busy devices
The USB or SD drive on some system may be powered off after the system
goes into the S3 suspend state. To avoid leaving a mount point in a stale
state while its associated physical drive is gone, the cros-disks client
on the Chrome side unmounts all the mount points it manages before the
system goes into suspend. However, an ongoing filesystem access may keep a
mount point busy, which is beyond the control of Chrome or cros-disks. We
used to force umount the mount points but that has become undesirable. For
instance, when force unmounting a mount point backed by a FUSE process,
umount2() reports an error and the mount point is left in a half-broken
state.

This CL changes how DiskManager handles unmounting of busy devices. The
force option instructed by Chrome is filtered out. Instead, DiskManager
first tries to perform a normal unmount. If that fails because of the
device being unmounted is busy, DiskManager retries with a lazy unmount
before giving up and reporting an error.

BUG=chromium:646224
TEST=Tested the following:
1. Build and run unit tests.
2. Run platform_CrosDisksDBus and platform_CrosDisksFilesystem.
3. Manually copy a big file to a mounted USB / SD drive, put the system in S3
   suspend before the copy operation completes, resume the system and observe
   that the USB / SD drive is remounted correctly by Files.app. Also observe
   /var/log/message to confirm that cros-disks retried a lazy unmount on the
   drive before the system went into suspend.

Change-Id: I51a645e8cb063b5a2725ce365d2a8428f4959474
Reviewed-on: https://chromium-review.googlesource.com/388592
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>