shill: remove member variable for max auto connection cooldown time in Service
This CL is a minor cleanup of CL:367500, which introduced a protected
member variable in the Service class for specifying the maximum auto
connect cooldown time that can be overwritten by a Service subclass like
CellularService. C++ style guide requires all non static const data
members to be private. But instead of simply making the variable
private, this CL replaces the variable with an overridable protected
const method Service::GetMaxAutoConnectCooldownTimeMilliseconds().

The rationale is that, in the current implementation, the maximum is
effectively a constant for a service type and isn't meant to be
adjustable through the lifetime of a service, so it's less desirable to
tax each service instance with an additional instance variable for that.

BUG=None
TEST=Unit tests.

Change-Id: Icaf15473d0829317b8a6674595705e72b6157fc1
Reviewed-on: https://chromium-review.googlesource.com/388754
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>