debugd: capture_utility: fix handling of non-wifi interfaces
We were sending the output of iw through a pipe which killed the exit
code and caused the func to always return true.  This in turn broke
the caller who tried to figure out if it's a WiFi interface.  This in
turn caused an error to be emitted which broke the success logic.

Fix the original func so now we can capture packets over eth0.

BUG=None
TEST=`packet_capture --device eth0` now works

Change-Id: I7ebe05cad8e4d2eb52045d5b57cc15635ebc6c42
Reviewed-on: https://chromium-review.googlesource.com/393087
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>