touch_keyboard: Add touch_keyboard module
This CL adds in the touch_keyboard module with contains the
low-level utilities for making a touch keyboard/touchpad combo
work. This patch adds the initial module setup such as build
scripts and several supporting libraries that are required to
get the eventual functionality.  For the sake of splitting up
reviews, the more complex "FakeKeyboard" library has been omitted
from this first CL, in favor of the simpler "FakeTouchpad."  Both
will be included eventually, but the touchpad is already working
well, so we might as well get it in while work continues on the
keyboard portion.  EvdevSource and UinputDevice form the basis
of both the FakeKeyboard and FakeTouchpad classes, so they must
be included first as well.

This first CL includes several classes that don't work/make sense
without each other:

FakeTouchpad:
  This class is a high-level abstraction that handles everything
  required to create a "fake" touchpad from a subregion of a larger
  touch device by using the other helper classes defined here.

EvdevSource:
  This is a class that can be derived from (as FakeTouchpad does) to
  easily allow an object to collect and process Evdev events from
  an input device.

UinputDevice:
  This is a class that can be derived from (as FakeTouchpad does) to
  easily allow an object to create a brand new input device for the
  OS by using the uinput module, and then generate input event
  programmatically.

MtStateMachine:
  This class processes Evdev events into a more manageable format
  by mimicing the kernel's state machine, and allows us to make sense
  of the events captured from the source touch sensor.

A patch will follow that adds the FakeKeyboard class in, using these.

The binary produced is called touch_keyboard_handler and is run at
boot by an upstart script that calls it through minijail as the
"touch_keyboard" user.

BUG=chrome-os-partner:59138
TEST=emerged and deployed on a pbody.  When I manually run the
generated executable I am able to move the pointer with the touchpad
portion of the larger touch surface.
CQ-DEPEND=CL:390506,CL:428351

Change-Id: Iae703e319731ff4f6d11682f7285e1aa5bd02aaf
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/400099
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>