container_utils: add device_jail FUSE daemon
This FUSE daemon exports a filesystem that can be mounted over
the /dev directory of a container. It should hide devices that
have not been shared with containers based on the pid of the
calling process, and expose a set of base devices as well as
spawn instances of device_jail for whitelisted devices.

It also passes through most files, but hides broken symlinks
pointing inside of the filesystem to avoid leaking the presence
of devices containers are not supposed to be able to see.

CQ-DEPEND=CL:406627,CL:410061
BUG=chromium:644338
TEST=attach e.g. android device; use lsusb -t to find what bus
  and device it is, then (for example):
  # start device-jail
  # cat /tmp/djfs/bus/usb/001/004
  and verify that we attempt to open /dev/jailed-189-3 via
  syslog, and that the request is routed through permission_broker,
  and that it's successful and the read succeeds;
  check that other devices are hidden

Change-Id: I5539c90fc8afa20d125afba03ae6a03e6ee20126
Reviewed-on: https://chromium-review.googlesource.com/406509
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>