login: UMA metrics VPD update status.
The result of VPD update is sent with UMA metrics as a bool value.
True when VPD update succeeded and false when failed. This will
allow us to have statistics on what devices are affected. The
metrics are submitted only for enrolled devices, when a flag in VPD
has to be changed.

BUG=chromium:653814
TEST=None

Change-Id: I32b9a01bf132ee914504ffa2a2b817f1d9004786
Reviewed-on: https://chromium-review.googlesource.com/400881
Commit-Ready: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Thiemo Nagel <tnagel@chromium.org>