gobi-cromo-plugin: fix buffers for retrieving radio interfaces
The Gobi SDK provides two library functions, GetDeviceCapabilities and
GetServingNetwork, which return an array of radio interfaces via a
buffer provided by the caller. These functions take the buffer as a
BYTE* argument and a buffer size arugment that specifies the maximum
number of elements that the buffer can hold. However, these functions
internally treat each element as ULONG and iterate through the buffer as
if it's a ULONG array.

When calling the two functions, the gobi-cromo-plugin code incorrectly
treat the radio interface buffer as BYTE* instead of ULONG* such that
the buffer size indicates more elements than the buffer can hold. This
could potentially overrun the buffer.

This CL makes minor changes to the gobi-cromo-plugin code to address the
issue. The gobi-cromo-plugin code would probably need more cleanups.

BUG=chromium:659432
TEST=`USE='cellular gobi' FEATURES=test emerge-$BOARD gobi-cromo-plugin`
TEST=Run network_3GModemControl with a Gobi3k modem.

Change-Id: I69a8833ff321a15f090586e5a08cf8e41313d9ce
Reviewed-on: https://chromium-review.googlesource.com/403513
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>