attestation: Don't put SetSystemSalt on a work thread
cryptohomed sends SetSystemSalt to attestationd as a part of its
initialization. If attestationd is busy with some long operation,
and SetSystemSalt is put as a task on a work thread, it has to wait
until that operation is over.
Since attestationd indeed typically runs PrepareForEnrollment that
can last for tens of seconds at this time, cryptohomed initialization
waits for attestation-related activities before it can accept
user mount related requests.
To avoid this priority inversion, process the small SetSystemSalt
operation right on the dbus thread.

BUG=chrome-os-partner:57422
TEST=login, watch how dbus commands are processed by attestationd
     (more details on the behavior before the change is provided in BUG)
     unit tests

Change-Id: I2ab641599cc0a36b7618923f3ecf3c8500847a56
Reviewed-on: https://chromium-review.googlesource.com/404308
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>