shill: fix declarations of string constant
String constants should be declared as 'const char* const' or
'const char []', but not 'const char*'. A 'const char*' variable is a
mutable pointer to a constant string literal, so the variable can be
changed to point to a different constant string literal, which isn't
what we typically expect for a string constant.

BUG=None
TEST=Build and run unit tests.

Change-Id: Icd53e3ef97a0cb3b745452110727949e7f758809
Reviewed-on: https://chromium-review.googlesource.com/404532
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>