crash: Avoid uploading empty reports
crash_sender gets the path of the next report to upload, goes to sleep
for rate limiting, and proceeds with uploading after waking up. If the
report is in the cryptohome and the user logs out in the meantime, it
uploads garbage. This CL adds error checking to prevent that.

BUG=b:31248243
TEST=Comment out check that disables sleeping for mock sending. Run
     mock crash_sender, log out, and observe that crash_sender does
     not upload inaccessible reports in ~chronos/crash.

Change-Id: I2adfc59173df5e64d9704bf077c2bf810be3d959
Reviewed-on: https://chromium-review.googlesource.com/405791
Commit-Ready: Dominik Laskowski <domlaskowski@google.com>
Tested-by: Dominik Laskowski <domlaskowski@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>