init: sanity check /var/lib too
A lot of daemons use /var/lib for state to keep across boots, so add it
to the list of paths to sanity check so daemons don't need to do it.

The misc subdir is used for one-off files and is a standard path too.

BUG=None
TEST=precq passes

Change-Id: Ib0b0d46f37a648c56c9beda8e0b33b43d3c4e9f4
Reviewed-on: https://chromium-review.googlesource.com/407027
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Ricky Zhou <rickyz@chromium.org>