power: tag internal stylus devices
We want to be able to disable wakeups from built-in stylus devices when
unit's lid is closed and user extracts the stylus, so we need to tag
such devices as "internal_stylus".

Udev does not make distinction between a tablet device and a stylus
device that complements touchscreen, but if we find a "tablet" connected
to an internal bus, such as i2c, on a Chromebook, we can assume it is
such stylus.

BUG=chrome-os-partner:59360
TEST=Try removing stylus on sleeping unit, then check logs to ensure we
did not wake up when we should not have.

Change-Id: I908068bbc4e55272291a7b20543498fbbf2a7b24
Reviewed-on: https://chromium-review.googlesource.com/408793
Commit-Ready: Dmitry Torokhov <dtor@chromium.org>
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>