login: Allow enrollment when VPD fails for parrot/glimmer devices.
Fixed the code that checks whether the device has a glimmer or
parrot board. Previous implementation was checking if the board
name matches exactly and it turns out there are some affected
devices with board name 'glimmer-signed-mp-v4keys'.

New implementation strips the board name removing the "-signed-.."
part and checking if the remaining string (the beginning part) is
the right one.

This implementation allows the parrot/glimmer devices to be enrolled
even if VPD update failed, thus limiting the number of users affected
by the bug (crbug/653814).

BUG=chromium:653814
TEST=Manually tested on a glimmer device.

Change-Id: Ic3953e123a5d70526a5096cae2a09db7d9c6d443
Signed-off-by: Igor <igorcov@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/408160
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Thiemo Nagel <tnagel@chromium.org>