biod: send biometric signals directly from the biometric object
Prior to this change, signals were sent from the biometric session's
D-Bus objects. However, this introduced a race condition between a
session being created by the client and that client connecting to the
signals. A signal could be sent after session creation but before the
client had time to connect to the signal.

With this change, the biometric object itself sends the signals, which
means the clients can subscribe to the signals before getting their
hands on the session object.

TEST=None
BUG=None

Change-Id: I203722fdc80e04ec5f6e06b88724aa1c4d957723
Reviewed-on: https://chromium-review.googlesource.com/410262
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>