shill: allow users of EventDispatcher to provide location with task
Since the location is provided inside EventDispatcher::PostTask and
EventDispatcher::PostDelayedTask, we often lose information about
where the task is actually coming from in minidumps and crash reports.
If we move the creation of the location object to the callsites of
EventDispatcher we should be able to get better information about
this.

BUG=None
TEST=unit tests

Change-Id: I405ad5de4e71005cbaf9d6962a1d72e1d37060f7
Reviewed-on: https://chromium-review.googlesource.com/411281
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>