installer: Ensure that the squashfs doesn't written to during install
MakeFileSystemRw only works on filesystems which can be made
writable, but doesn't check ensure that the filesystem is writable at all.
This change checks that if the filesystem has the squashfs
id at the beginning, and if so, does not modify it at all.

Also, it removes the unused "rw" argument from "MakeFileSystemRw"
Given its name, it should always make the filesystem rw,
and is currently always used that way.

BUG=chromium:666530

TEST=I did good things.

Change-Id: I2fc0b8304468aab2757593a56f50e1f0a41fe3b1
Reviewed-on: https://chromium-review.googlesource.com/412271
Commit-Ready: Matthew Mirman <mirman@chromium.org>
Tested-by: Matthew Mirman <mirman@chromium.org>
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>