login: Request for state keys done only after the time sync completed.
This prevents the state keys to be generated using a wrong timestamp.

Because we can't get a notification from tlsdated when the time gets
updated (crbug/661205), this implementation is making requests for
LastTimeSyncInfo in loops until a response that time sync was done is
received.

BUG=chromium:649422
TEST=manually tested on a reks device.

Change-Id: I0a91e338a211ff76e23f9e2080ab9c62deeb3225
Reviewed-on: https://chromium-review.googlesource.com/406948
Commit-Ready: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>