run_oci: Add options for cgroup parent
Add an option for setting the parent cgroup directory.
This option will be used to specify the user_containers subdirectory of
each cgroup.  This directory will be owned by the containers user so
that subdirectories for each app can be created within that while the
system can decide how many resources the containers get as one aggregate
entity.

BUG=665622
TEST=Manual: set run_oci with and without directory specified
security_RunOci

Change-Id: Idb02a58f4d8c0ce5551caa7ae08fbcbeade0a1b0
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/414215
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>