crash: Handle new warning format
The kernel has switched its warning format to include CPU and PID in
the WARNING line per [1] which landed in the 3.11 cycle. Kernels 3.10
and earlier continue to use the old format. Adjust the analyzer to
understand both old and new format. Also update the test to use a
warning message in new format for testing most edge case behavior, but
keep a copy of the old format and make sure it still gets picked up by
the collector.

[1] https://kernel.googlesource.com/pub/scm/linux/kernel/git/torvalds/linux.git/+/dcb6b45254e2281b6f99ea7f2d51343954aa3ba8

BUG=chromium:668134
TEST=FEATURES=test emerge-$BOARD -v1 crash-reporter

Change-Id: I41617d1a112170961d198625caba58d35a5ef385
Reviewed-on: https://chromium-review.googlesource.com/414905
Commit-Ready: Mattias Nissler <mnissler@chromium.org>
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>