vpn: Extend l2tp/ppp timeout from 10s to 60s
The existing logic assumes that l2tp+ppp will finish connecting within
10 seconds of the ESP SA being established.  This is fairly aggressive,
and it does not cope well with slow links / packet loss / buggy gateways.

Enable redial in xl2tpd so that it can retry a couple of times if the
initial attempt fails, and extend the default ppp_setup_timeout to 60s
(as shill does not override this parameter).

BUG=chromium:619273
TEST=connect to a "slow" server

Change-Id: I2a14766e349e6a99763760bce47207120fd11d30
Reviewed-on: https://chromium-review.googlesource.com/414925
Commit-Ready: Kevin Cernekee <cernekee@chromium.org>
Tested-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>