shill: Fix L2TP certificate metrics
Commit c350e68360d0 ("Further deprecate CaCertNSS") mistakenly removed
all accounting for Network.Shill.Vpn.RemoteAuthenticationType
L2TP/IPSec Certificate, so it shows as 0 in UMA.  While certificate
auth remains a remarkably unpopular option (by a factor of ~3000:1 in
UserAuthenticationType) it shouldn't be zero.  Revert that part of
the commit, and update the code to look for a PemProperty instead of
an NSS handle.

BUG=chromium:486182
TEST=add debug prints in each clause, then connect to PSK + cert VPNs
     to verify that the check is WAI
TEST=`FEATURES=test emerge-link shill`

Change-Id: I3cfb050e45f3a1b38fbd3927b0828445273de3a7
Reviewed-on: https://chromium-review.googlesource.com/414439
Commit-Ready: Kevin Cernekee <cernekee@chromium.org>
Tested-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>