trunks: optional hash generation in Sign() and Verify()
TpmUtility doesn't always need to generate the hash of a plaintext in Sign()
and Verify(): sometimes a ready-to-use hash of the input data is generated by
the called and passed to these methods. Chaps does that, for example.
To support such use cases, add a parameter to Sign() and Verify() that tells
if the hash needs to be generated by TpmUtility.

CQ-DEPEND=CL:414956
BUG=chrome-os-partner:59754
TEST=1) For a pre-generated key pair (/tmp/priv.key + /tmp/pub.key) and input
     /tmp/1.txt, import the private key, sign the text using it, and verify
     that the signature is correct:
       openssl pkcs8 -inform pem -outform der -in /tmp/priv.key \
                     -out /tmp/priv.der -nocrypt
       p11_replay --import --path=/tmp/priv.der --type=privkey -id=bbbbbb
       pkcs11-tool --module=`ls /usr/lib**/libchaps.so` --slot=0 --id=bbbbbb \
                   --sign </tmp/1.txt >/tmp/1.sig
       openssl dgst -sha1 -verify /tmp/pub.key -signature /tmp/1.sig /tmp/1.txt
     The last operation should say "Verified OK".
     2) Unit tests.

Change-Id: Ia88233408af6d3e2bd08de1fa534b6507c619945
Reviewed-on: https://chromium-review.googlesource.com/414278
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>